last updates

+++ src/blog/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.http import HttpResponse
-from .models import Post
+from .models import Post, Like
 from .forms import PostForm, CommentForm
 
 # Create your views here.
@@ -52,6 +52,8 @@ def post_detail(request, slug):
 
 def post_update(request, slug):
     obj = get_object_or_404(Post, slug=slug)
+    if request.user.id != obj.author.id:
+        return HttpResponse("You are not authorized!!")
     form = PostForm(request.POST or None, request.FILES or None, instance=obj)
     if form.is_valid():
         form.save()
@@ -63,6 +65,9 @@ def post_update(request, slug):
 
 def post_delete(request, slug):
     obj = get_object_or_404(Post, slug=slug)
+    
+    if request.user.id != obj.author.id:
+        return HttpResponse("You are not authorized!!")
     if request.method == "POST":
         obj.delete()
         return redirect("blog:list")
@@ -70,7 +75,15 @@ def post_delete(request, slug):
         "object" : obj
     }
     return render(request, "blog/post_delete.html", context)
-
+  
 def like(request, slug):
     if request.method == "POST":
-        obj = get_object_or_404()
+        obj = get_object_or_404(Post, slug=slug)
+        like_qs = Like.objects.filter(user=request.user, post=obj)
+        if like_qs:
+            like_qs.delete()
+        else:
+            Like.objects.create(user=request.user, post=obj)
+        return redirect('blog:detail', slug=slug)
+    return redirect('blog:detail', slug=slug)
+
